refactor(components): move layout components to src/components/layout
