+++ frontend/src/components/layout/sidebar/Sidebar.scss
