+++ frontend/src/components/layout/sidebar/sidebar-header/SidebarHeader.tsx
