+++ frontend/src/components/layout/topbar/Topbar.scss
